requirements/req.txt
python-dotenv==1.1.0
requests==2.32.3
PyGithub==2.6.1
cryptography==44.0.2
colorama==0.4.6
tqdm==4.67.1
rich==14.0.0
pytz==2025.2
flask==3.1.0